examples/spark-data-lake/infra/requirements-dev.txt
# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
aws-cdk.lambda-layer-kubectl-v30
black
cdk-nag
pytest
pytest==7.4.3
